WIP move controller_authorized decorator from auth package
auth package should contain only 3bot related authetication helpers for now
